requirements.txt
opencv-python
mediapipe-silicon